+++ lib/plausible_web/controllers/stats_controller.ex
@@ -157,6 +157,9 @@ defmodule PlausibleWeb.StatsController do
         ~c"browsers.csv" => fn -> Api.StatsController.browsers(conn, params) end,
         ~c"browser_versions.csv" => fn -> Api.StatsController.browser_versions(conn, params) end,
         ~c"operating_systems.csv" => fn -> Api.StatsController.operating_systems(conn, params) end,
+        ~c"operating_system_versions.csv" => fn ->
+          Api.StatsController.operating_system_versions(conn, params)
+        end,
         ~c"devices.csv" => fn -> Api.StatsController.screen_sizes(conn, params) end,
         ~c"conversions.csv" => fn -> Api.StatsController.conversions(conn, params) end,
         ~c"referrers.csv" => fn -> Api.StatsController.referrers(conn, params) end,